requirements.txt
pathlib~=1.0.1
cffi~=1.15.1
defusedxml~=0.7.1
typing_extensions~=4.3.0
future~=0.18.2
pillow~=10.0.1
ipython~=7.31.1
zlib~=1.2.13
olefile~=0.46
jupyter~=1.0.0
pip~=22.2.2
panel~=0.13.1
dataclasses~=0.8
attrs~=21.4.0
wheel~=0.37.1
openssl~=3.0.13
cryptography~=39.0.1
docutils~=0.18.1
sphinx~=5.0.2
tornado~=6.1
jinja2~=3.1.2
keyring~=23.4.0
tomli~=2.0.1
importlib_metadata~=4.11.3
markupsafe~=2.1.5
werkzeug~=2.2.2
click~=8.1.7
blinker~=1.7.0
itsdangerous~=2.1.2
setuptools~=68.2.2
visitor~=0.1.3
chardet~=4.0.0
wtforms~=3.1.2
dominate~=2.9.1
greenlet~=1.1.1
psutil~=5.9.0
flask~=3.0.2
babel~=2.9.1
reportlab~=4.0.9
packaging~=23.2
zipp~=3.8.0
cython~=0.29.32
sqlalchemy~=2.0.28
mock~=4.0.3
pytest~=7.1.2
pyodbc~=4.0.34
platformdirs~=2.5.2
requests~=2.28.1
yagmail~=0.15.293
mailersend~=0.5.6
python-dotenv~=0.21.0
gunicorn==21.2.0
psycopg2-binary==2.9.9